db_upgrade.py
#!python2
from migrate.versioning import api
from config import SQLALCHEMY_DATABASE_URI
from config import SQLALCHEMY_MIGRATE_REPO
api.upgrade(SQLALCHEMY_DATABASE_URI, SQLALCHEMY_MIGRATE_REPO)
print 'Current database version: ' + str(api.db_version(SQLALCHEMY_DATABASE_URI, SQLALCHEMY_MIGRATE_REPO))

'''
copy the migration scripts to your production server and run a simple script that applies the changes for you. 
The database upgrade can be done with this little Python script (file db_upgrade.py)

'''
